split typesarch and titlesearch into smaller components
