+++ src/style/searchBar.css
@@ -36,3 +36,11 @@ input[type="text"]:focus {
   background-color: #e6e6e6;
   color: #33314e;
 }
+#dropdown-button-dark{
+  height:95px;
+  width:100px;
+  background-color:#33314e;
+  outline:none;
+  border:none;
+  border-radius:16px
+}
